added image and placholder synopsis to movie pages (not yet updated css)
